+++ package.json
@@ -9,19 +9,22 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@ably-labs/react-hooks": "^2.0.9",
+    "@ably-labs/react-hooks": "^2.1.1",
     "@types/node": "18.11.9",
     "@types/react": "18.0.25",
     "@types/react-dom": "18.0.8",
-    "dotenv": "^16.0.3",
+    "dotenv": "^16.3.1",
     "eslint": "8.27.0",
     "eslint-config-next": "13.0.2",
-    "next": "13.0.2",
+    "next": "13.4.7",
     "react": "18.2.0",
     "react-dom": "18.2.0",
     "typescript": "4.8.4"
   },
   "devDependencies": {
-    "eslint-plugin-react-hooks": "^4.6.0"
+    "autoprefixer": "^10.4.14",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "postcss": "^8.4.24",
+    "tailwindcss": "^3.3.2"
   }
 }
